app/src/main/java/com/ntl/todoapp/database/TaskDatabase.kt
package com.ntl.todoapp.database

import android.content.Context
import androidx.room.Database
import androidx.room.RoomDatabase
import androidx.room.Room
import com.ntl.todoapp.model.Task

@Database(entities = [Task::class], version = 1)
abstract class TaskDatabase : RoomDatabase() {

    abstract fun getTaskDAO(): TaskDAO

    companion object {
        private const val DATABASE_NAME = "task.db"

        private var instance: TaskDatabase? = null

        @JvmStatic
        @Synchronized
        fun getInstance(context: Context): TaskDatabase? {
            if (instance == null) {
                instance = Room.databaseBuilder(
                    context.applicationContext,
                    TaskDatabase::class.java,
                    DATABASE_NAME
                )
                    .allowMainThreadQueries()
                    .build()
            }
            return instance
        }
    }
}